requirements.txt
requests>=2.31
psycopg2-binary>=2.9
python-dotenv>=1.0
streamlit>=1.35
plotly>=5.20
pandas>=2.1
